♻️ refactor: react query 코드 api/queries 폴더 밑으로 변경
